+++ assets/js/welcome/index.js
@@ -4,6 +4,8 @@ import page from "page";
 import pages from "./pages";
 import $ from "jquery";
 import "bootstrap";
+import AOS from "aos";
+import slideshow from "./components/slideshow";
 
 log("Welcome to NordicSoft Express 4 Template! Environment: " + options.env);
 
@@ -25,3 +27,9 @@ $("body").tooltip({
 $.ajaxSetup({
     cache: false,
 });
+
+// setup AOS
+AOS.init({ once: true });
+
+// setup Slick
+slideshow($("header .slideshow"));